requirements-dev.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pandas>=2.0.0
pyyaml>=6.0.0
